100l, SIZE_MAX must be UINT_MAX to prevent an integer overflow later on, also it is preferable if demuxer parsing does not differ between architectures.
git-svn-id: svn://svn.mplayerhq.hu/mplayer/trunk@29640 b3059339-0415-0410-9bf9-f77b7e298cf2

+++ libmpdemux/demux_mov.c
@@ -1445,7 +1445,7 @@ static void lschunks(demuxer_t* demuxer,int level,off_t endpos,mov_track_t* trak
 	    z_stream zstrm;
 	    stream_t* backup;
 
-	    if (moov_sz > SIZE_MAX - 16) {
+	    if (moov_sz > UINT_MAX - 16) {
               mp_msg(MSGT_DEMUX, MSGL_ERR, "Invalid cmvd atom size %d\n", moov_sz);
               break;
             }
